renaming folder, adding spinne with tailwind css
